Moving Api to delete Name Type from studies to studyService and StudyResource
Issue: IBP-5778/IBP-5983
Review: none
