+++ src/main/java/org/ibp/api/java/dataset/DatasetService.java
@@ -352,5 +352,4 @@ public interface DatasetService {
 
 	List<GermplasmNameTypeDTO> getAllPlotDatasetNames(Integer datasetId);
 
-	void deleteNameTypeFromStudies(Integer nameTypeId);
 }
